# docs/swagger.yaml
basePath: /api
definitions:
  handlers.CreateShiftSummaryRequest:
    properties:
      generation_summaries:
        items:
          $ref: '#/definitions/handlers.GenerationSummaryRequest'
        type: array
      session_id:
        type: string
      shift_note:
        type: string
      summary_date:
        type: string
      summary_time:
        type: string
    type: object
  handlers.GenerationStationListResponse:
    properties:
      station_id:
        type: string
      station_name:
        type: string
    type: object
  handlers.GenerationStationResponse:
    properties:
      reserve_energy_mw:
        type: number
      running_units:
        type: integer
      station_id:
        type: string
      station_name:
        type: string
    type: object
  handlers.GenerationSummaryRequest:
    properties:
      reserve_energy_mw:
        type: number
      running_units:
        type: integer
      station_id:
        type: string
    type: object
  handlers.ShiftSummaryResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      created_by_name:
        type: string
      generation_stations:
        items:
          $ref: '#/definitions/handlers.GenerationStationResponse'
        type: array
      id:
        type: string
      session_id:
        type: string
      shift_note:
        type: string
      summary_date:
        type: string
      summary_time:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A REST API for E-Logging with authentication, device management, and
    log tracking.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Logging API
  version: "1.0"
paths:
...
  /shift-summary:
    post:
      consumes:
      - application/json
      description: Create a shift summary with generation station running units and
        reserve energy
      parameters:
      - description: Shift summary details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateShiftSummaryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ShiftSummaryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create shift summary
      tags:
      - Shift Summary
  /shift-summary/{id}:
    get:
      consumes:
      - application/json
      description: Get a shift summary with all generation station details
      parameters:
      - description: Summary ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ShiftSummaryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get shift summary
      tags:
      - Shift Summary
  /shift-summary/session/{sessionId}:
    get:
      consumes:
      - application/json
      description: Get the shift summary associated with a specific operator session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ShiftSummaryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get session shift summary
      tags:
      - Shift Summary
  /shift-summary/stations/generation:
    get:
      consumes:
      - application/json
      description: Get all stations with type "Generation" for shift summary creation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.GenerationStationListResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get generation stations
      tags:
      - Shift Summary
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"